feat(deleteCave): move merge caves code into delete route
