+++ api/services/CaveService.js
@@ -1,3 +1,5 @@
+const ramda = require('ramda');
+
 module.exports = {
   /**
    * @param {Object} caves caves to set
@@ -54,6 +56,96 @@ module.exports = {
         : undefined;
 
       return createdCave;
+    });
+  },
+
+  /**
+   * Merge a source cave into a destination cave (no checks performed)
+   * @param {*} req
+   * @param {*} res
+   * @param {*} sourceCaveId cave id to be merged into destinationCave
+   * @param {*} destinationCaveId cave id which will receive sourceCaveId data
+   * @throws Sails ORM errors (see https://sailsjs.com/documentation/concepts/models-and-orm/errors)
+   */
+  mergeCaves: async (sourceCaveId, destinationCaveId) => {
+    await sails.getDatastore().transaction(async (db) => {
+      // Delete sourceCave names
+      await TName.destroy({ cave: sourceCaveId }).usingConnection(db);
+
+      // Move associated data
+      await TComment.update({ cave: sourceCaveId })
+        .set({ cave: destinationCaveId })
+        .usingConnection(db);
+      await TDescription.update({ cave: sourceCaveId })
+        .set({ cave: destinationCaveId })
+        .usingConnection(db);
+      await TDocument.update({ cave: sourceCaveId })
+        .set({ cave: destinationCaveId })
+        .usingConnection(db);
+      await TEntrance.update({ cave: sourceCaveId })
+        .set({ cave: destinationCaveId })
+        .usingConnection(db);
+      await THistory.update({ cave: sourceCaveId })
+        .set({ cave: destinationCaveId })
+        .usingConnection(db);
+
+      // Handle many-to-many relationships
+      const sourceCave = await TCave.findOne(sourceCaveId)
+        .populate('exploringGrottos')
+        .populate('partneringGrottos')
+        .usingConnection(db);
+      const destinationCave = await TCave.findOne(destinationCaveId)
+        .populate('exploringGrottos')
+        .populate('partneringGrottos')
+        .usingConnection(db);
+
+      const {
+        exploringGrottos: sourceExplorers,
+        partneringGrottos: sourcePartners,
+      } = sourceCave;
+      const {
+        exploringGrottos: destinationExplorers,
+        partneringGrottos: destinationPartners,
+      } = destinationCave;
+
+      // Update explored / partnered caves only if not already explored / partnered by the destination cave.
+      for (const sourceExp of sourceExplorers) {
+        if (!destinationExplorers.some((g) => g.id === sourceExp.id)) {
+          await TCave.addToCollection(
+            destinationCaveId,
+            'exploringGrottos',
+            sourceExp.id,
+          ).usingConnection(db);
+        }
+      }
+      for (const sourcePartn of sourcePartners) {
+        if (!destinationPartners.some((g) => g.id === sourcePartn.id)) {
+          await TCave.addToCollection(
+            destinationCaveId,
+            'partneringGrottos',
+            sourcePartn.id,
+          ).usingConnection(db);
+        }
+      }
+
+      // Update cave data with destination data being prioritised
+      const mergedData = ramda.mergeWith(
+        (a, b) => (b === null ? a : b),
+        sourceCave,
+        destinationCave,
+      );
+
+      const {
+        id,
+        exploringGrottos,
+        partneringGrottos,
+        ...cleanedMergedData
+      } = mergedData;
+      await TCave.update(destinationCaveId)
+        .set(cleanedMergedData)
+        .usingConnection(db);
+
+      await TCave.destroy(sourceCaveId).usingConnection(db);
     });
   },
 };
