+++ config/policies.js
@@ -35,9 +35,8 @@ module.exports.policies = {
     addDocument: 'tokenAuth',
     setMassif: 'tokenAuth',
     create: 'tokenAuth',
-    delete: ['tokenAuth', 'moderatorAuth'],
+    delete: 'tokenAuth',
     update: 'tokenAuth',
-    merge: 'tokenAuth',
   },
 
   'v1/CaverController': {
